3D Shape Manipulator
This is another project that I created for my Graphics class. What the
program does is take a 3D shape and manipulates it. Some of these
manipulations include rotation, scaling, and translation. Please note
that the UI in this version unfortunately did not turn out very well in
this version.